fix: prevent re-backup symlink on Unix system

+++ scripts/fish/enable.sh
@@ -4,7 +4,7 @@ FISH_SOURCE=$PWD/fish
 FISH_DESTINATION=~/.config/fish
 FISH_BACKUP=~/.config/fish_backup
 
-if [ -d "$FISH_DESTINATION" ]; then
+if [ -d "$FISH_DESTINATION" ] && [ ! -L "$FISH_DESTINATION" ]; then
 	echo "Creating backup directory for fish..."
 	mv $FISH_DESTINATION $FISH_BACKUP
 fi
